+++ src/lib.rs
@@ -206,46 +206,65 @@ pub mod _priv {
     unsafe impl Send for __CFString {}
     unsafe impl Sync for __CFString {}
 
-    /// Returns `true` if the string is a valid C string.
-    pub const fn is_cstr(s: &str) -> bool {
-        if s.is_empty() {
-            return false;
-        }
-
-        let bytes = s.as_bytes();
-        let mut i = 0;
-        let last_i = bytes.len() - 1;
-        loop {
-            if bytes[i] == 0 {
-                return i == last_i;
-            }
-
-            i += 1;
-
-            if i > last_i {
-                return false;
+    pub mod str {
+        /// Returns `true` if `s` ends with a 0 byte.
+        pub const fn is_nul_terminated(s: &str) -> bool {
+            match s.as_bytes() {
+                [.., 0] => true,
+                _ => false,
             }
         }
-    }
 
-    #[cfg(test)]
-    mod tests {
+        /// Returns `true` if `s` contains only non-null ASCII bytes.
+        pub const fn is_cf_ascii(s: &str) -> bool {
+            // Remove the trailing null byte if there is one.
+            let bytes = match s.as_bytes() {
+                [b @ .., 0] => b,
+                b => b,
+            };
 
-        #[test]
-        fn is_cstr() {
-            let cases: &[_] = &[
-                ("\0", true),
-                ("\0\0", false),
-                ("\0\0\0", false),
-                ("", false),
-                ("a", false),
-                ("a\0", true),
-                ("\0a\0", false),
-                ("a\0\0", false),
-            ];
+            let mut i = 0;
+            loop {
+                if i == bytes.len() {
+                    return true;
+                }
 
-            for &(s, is_cstr) in cases {
-                assert_eq!(super::is_cstr(s), is_cstr, "invalid result for {:?}", s);
+                let byte = bytes[i];
+                if !byte.is_ascii() || byte == 0 {
+                    return false;
+                }
+
+                i += 1;
+            }
+        }
+
+        #[cfg(test)]
+        mod tests {
+
+            #[test]
+            fn is_cf_ascii() {
+                let cases: &[_] = &[
+                    ("\0", true),
+                    ("\0\0", false),
+                    ("\0\0\0", false),
+                    ("", true),
+                    ("a", true),
+                    ("ä", false),
+                    ("a\0", true),
+                    ("\0a\0", false),
+                    ("a\0\0", false),
+                    ("\0ä\0", false),
+                    ("ä\0\0", false),
+                ];
+
+                for &(s, is_cf_ascii) in cases {
+                    assert_eq!(
+                        super::is_cf_ascii(s),
+                        is_cf_ascii,
+                        "invalid result for {:?}",
+                        s
+                    );
+                }
             }
         }
     }