+++ CHANGELOG.md
@@ -7,10 +7,27 @@ The format is based on [Keep a Changelog] and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- The `nsstring!` macro can now take `const X: &str` as input, not just string
+  literals.
+
+- The `nsstring!` macro now allows trailing null bytes and uses the constant
+  as-is.
+
+  This makes it possible for input data to not get emitted in the binary twice.
+  The compiler currently does not coalesce string prefixes and instead emits the
+  same prefix data twice.
+
 ### Changed
 
 - **\[breaking\]** Increased crate `#[cfg]` strictness from any 32/64 bit to
   only target x86 and ARM.
+
+- **\[breaking\]**  The `nsstring!` macro can only take ASCII strings.
+
+  The canonical Unicode representation is UTF-16, so any non-ASCII strings must
+  be transcoded to UTF-16. See issue [#3].
 
 ## [0.2.0] - 2020-09-11
 
@@ -39,6 +56,7 @@ Initial release.
 [Keep a Changelog]:    http://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: http://semver.org/spec/v2.0.0.html
 
+[#3]: https://github.com/nvzqz/fruity/issues/3
 [#1]: https://github.com/nvzqz/fruity/issues/1
 
 [Unreleased]: https://github.com/nvzqz/fruity/compare/v0.2.0...HEAD